+++ pychunkedgraph/meshing/mesh_analysis.py
@@ -1,5 +1,53 @@
 from cloudvolume import CloudVolume, Storage
+import numpy as np
+import os
 from pychunkedgraph.meshing import meshgen, meshgen_utils
+
+
+def compute_centroid_by_range(vertices):
+    bbox_min = np.amin(vertices, axis=0)
+    bbox_max = np.amax(vertices, axis=0)
+    return bbox_min + ((bbox_max - bbox_min) / 2)
+
+
+def compute_centroid_with_chunk_boundary(cg, vertices, l2_id, last_l2_id):
+    """
+    Given a level 2 id, the vertices of its mesh, and the level 2 id preceding it in
+    a path, return the center point of the mesh on the chunk boundary separating the two
+    ids, and the center point of the entire mesh.
+    :param cg: ChunkedGraph object
+    :param vertices: [[np.float]]
+    :param l2_id: np.uint64
+    :param last_l2_id: np.uint64 or None
+    :return: [np.float]
+    """
+    centroid_by_range = compute_centroid_by_range(vertices)
+    if last_l2_id is None:
+        return [centroid_by_range]
+    l2_id_cc = cg.get_chunk_coordinates(l2_id)
+    last_l2_id_cc = cg.get_chunk_coordinates(last_l2_id)
+
+    # Given the coordinates of the two level 2 ids, find the chunk boundary
+    axis_change = 2
+    look_for_max = True
+    if l2_id_cc[0] != last_l2_id_cc[0]:
+        axis_change = 0
+    elif l2_id_cc[1] != last_l2_id_cc[1]:
+        axis_change = 1
+    if np.sum(l2_id_cc - last_l2_id_cc) > 0:
+        look_for_max = False
+    if look_for_max:
+        value_to_filter = np.amax(vertices[:, axis_change])
+    else:
+        value_to_filter = np.amin(vertices[:, axis_change])
+    chunk_boundary_vertices = vertices[
+        np.where(vertices[:, axis_change] == value_to_filter)
+    ]
+
+    # Get the center point of the mesh on the chunk boundary
+    bbox_min = np.amin(chunk_boundary_vertices, axis=0)
+    bbox_max = np.amax(chunk_boundary_vertices, axis=0)
+    return [bbox_min + ((bbox_max - bbox_min) / 2), centroid_by_range]
 
 
 def compute_mesh_centroids_of_l2_ids(cg, l2_ids, flatten=False):
